ANI: Fix local refs allocation for huge nr_refs
Issue: #IBV3OO

+++ static_core/plugins/ets/runtime/mem/ets_reference.h
@@ -142,7 +142,7 @@ public:
         return EtsReference::CastFromReference(PopLocalFrame(EtsReference::CastToReference(result)));
     }
 
-    bool EnsureLocalEtsCapacity(size_t capacity)
+    bool EnsureLocalEtsCapacity(uint32_t capacity)
     {
         return EnsureLocalCapacity(capacity);
     }
